# perf: use node's async esm module loader above node>=12