--- a/packages/amplify-cli/bin/amplify
+++ b/packages/amplify-cli/bin/amplify
@@ -1,3 +1,6 @@
 #!/usr/bin/env node
-require = require('esm')(module, { cache: false });
+const semver = require('semver');
+if (semver.compare(process.version, "v12.0.0") < 0) {
+  require = require('esm')(module, { cache: false });
+}
 require('../lib/index').run();