--- a/packages/amplify-cli/scripts/post-install.js
+++ b/packages/amplify-cli/scripts/post-install.js
@@ -1,4 +1,7 @@
-require = require('esm')(module, { cache: false });
+const semver = require('semver');
+if (semver.compare(process.version, 'v12.0.0') < 0) {
+  require = require('esm')(module, { cache: false });
+}
 const chalk = require('chalk');
 const fs = require('fs-extra');
 const path = require('path');